+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Fix course run computed state. A run without end date must be ONGOING_OPEN
 - Fix cookiecutter circleci gitlint configuration
 - Opened Course Run show empty date as '...'
+- Fix dashboard enrollment listing when they're linked to a product certificate.
 
 ### Removed